Added ParticleAffector base class and a few subclasses
Nearly all particle system stuff is in place at this point, I just have to figure out why the texture alpha component is not being handled correctly.  Instead of the zero-alpha areas of a texture being transparent and the one-alpha areas being opaque (and getting the glColor() value), everything is opaque, and the zero-alpha areas get the glColor() value, while the one-alpha areas are full white (ie. get the full texture color value).