Temporary fix to particle texture blending bug
The problem is that Skybox generates its own textures instead of getting them through the ImageManager which uses SFML to generate them.  So Skybox alters the global texture environment mode to GL_DECAL instead of GL_MODULATE.  Skybox needs to be refactored to store references to sf::Image's, once this is done, other things that are being textured might need to be updated to reflect the change as well.