Fixed inline fiasco.
Since the bodies of the functions were defined outside the class
definition, the inline keyword _was_ needed after all. Otherwise
the function bodies are compiled in each source file in which the
header is included, violating the One Definition Rule and leading
to a slew of linker errors regarding multiple definitions.

This fix is essentially a revert of my previous commit, with one
minor change: a larger function (BoundingBox::intersect)
is placed in its source file instead of the header. My apologies
fracking up the repo and not even doing a smoke test to make sure
it built.